chore(ci): enforce clippy pedantic lints in CI

+++ contracts/sanitization/src/lib.rs
@@ -38,7 +38,6 @@ impl core::fmt::Display for SanitizationError {
 
 /// Validates a general-purpose string: non-empty, within `max_len` bytes,
 /// no null bytes, no ASCII control characters (allows tab/LF/CR).
-#[must_use]
 pub fn sanitize_string(_env: &Env, input: &String, max_len: u32) -> Result<(), SanitizationError> {
     let len = input.len();
 
@@ -46,7 +45,7 @@ pub fn sanitize_string(_env: &Env, input: &String, max_len: u32) -> Result<(), S
         return Err(SanitizationError::EmptyInput);
     }
     // Guard: reject inputs that would overflow the static buffer.
-    if len > BUFFER_SIZE as u32 || len > max_len {
+    if (len as usize) > BUFFER_SIZE || len > max_len {
         return Err(SanitizationError::InputTooLong);
     }
 
@@ -74,14 +73,13 @@ pub fn sanitize_string(_env: &Env, input: &String, max_len: u32) -> Result<(), S
 
 /// Validates a human name: letters (any UTF-8), digits, spaces, hyphens,
 /// apostrophes, commas, and periods only (ASCII subset).
-#[must_use]
 pub fn sanitize_name(_env: &Env, input: &String) -> Result<(), SanitizationError> {
     let len = input.len();
 
     if len == 0 {
         return Err(SanitizationError::EmptyInput);
     }
-    if len > BUFFER_SIZE as u32 || len > MAX_NAME_LEN {
+    if (len as usize) > BUFFER_SIZE || len > MAX_NAME_LEN {
         return Err(SanitizationError::InputTooLong);
     }
 
@@ -110,14 +108,13 @@ pub fn sanitize_name(_env: &Env, input: &String) -> Result<(), SanitizationError
 
 /// Validates an email address: single '@', non-empty local and domain parts,
 /// domain contains at least one '.', all chars from the RFC 5321 allowed set.
-#[must_use]
 pub fn sanitize_email(_env: &Env, input: &String) -> Result<(), SanitizationError> {
     let len = input.len();
 
     if len == 0 {
         return Err(SanitizationError::EmptyInput);
     }
-    if len > BUFFER_SIZE as u32 || len > MAX_EMAIL_LEN {
+    if (len as usize) > BUFFER_SIZE || len > MAX_EMAIL_LEN {
         return Err(SanitizationError::InputTooLong);
     }
 
@@ -171,14 +168,13 @@ pub fn sanitize_email(_env: &Env, input: &String) -> Result<(), SanitizationErro
 
 /// Validates an identifier: alphanumeric chars, hyphens, underscores, colons,
 /// dots, and forward slashes (covers DIDs, slugs, and resource paths).
-#[must_use]
 pub fn sanitize_id(_env: &Env, input: &String) -> Result<(), SanitizationError> {
     let len = input.len();
 
     if len == 0 {
         return Err(SanitizationError::EmptyInput);
     }
-    if len > BUFFER_SIZE as u32 || len > MAX_ID_LEN {
+    if (len as usize) > BUFFER_SIZE || len > MAX_ID_LEN {
         return Err(SanitizationError::InputTooLong);
     }
 
@@ -200,14 +196,13 @@ pub fn sanitize_id(_env: &Env, input: &String) -> Result<(), SanitizationError> 
 }
 
 /// Validates a URL: printable ASCII only, length within MAX_URL_LEN.
-#[must_use]
 pub fn sanitize_url(_env: &Env, input: &String) -> Result<(), SanitizationError> {
     let len = input.len();
 
     if len == 0 {
         return Err(SanitizationError::EmptyInput);
     }
-    if len > BUFFER_SIZE as u32 || len > MAX_URL_LEN {
+    if (len as usize) > BUFFER_SIZE || len > MAX_URL_LEN {
         return Err(SanitizationError::InputTooLong);
     }
